moved pattern to top of concise submodule
