+++ langchain/concise/pattern.py
